refactor(api_client): split into directory module with domain submodules
Convert api_client.rs (2224 LOC) into a directory module and extract
6 submodules:

- upload.rs: asset upload flow, checksum, retry, timezone fixup (261 LOC)
- albums.rs: album list, create, cache, add-assets-to-album (361 LOC)
- library.rs: fetch albums/thumbnails/details, download, delete (282 LOC)
- search.rs: smart/OCR/metadata search, people, places, stats (386 LOC)
- errors.rs: RequestContext, HTTP/network error classification (145 LOC)
- upload_helpers.rs: MIME, timezone, file timestamp helpers (129 LOC)

mod.rs retains the ImmichApiClient struct, settings management,
connectivity failover, and type definitions (731 LOC).

Split impl blocks use pub(super) for cross-submodule method calls.
All crate::api_client:: public imports remain stable. All 101 tests
pass, clippy clean.
